Fix firestore issue
