+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.registeractivity">
+    package="com.example.blooddonation">
 
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
     <uses-permission android:name="android.permission.INTERNET"/>
@@ -16,12 +16,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.RegisterActivity"
+        android:theme="@style/Theme.MainActivity"
         tools:targetApi="31">
         <activity
-            android:name=".RegisterActivity"
+            android:name="com.example.blooddonation.ui.MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.RegisterActivity">
+            android:theme="@style/Theme.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
