+++ app/src/main/java/com/example/blooddonation/ui/MainActivity.kt
@@ -1,0 +1,20 @@
+package com.example.blooddonation.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.google.firebase.FirebaseApp
+import com.google.firebase.auth.FirebaseAuth
+
+class MainActivity : ComponentActivity() {
+    private lateinit var auth: FirebaseAuth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        FirebaseApp.initializeApp(this)
+        auth = FirebaseAuth.getInstance()
+        setContent {
+            BloodBankApp(auth.currentUser)
+        }
+    }
+}
