+++ app/src/main/java/com/example/blooddonation/ui/dashboard/Dashboard.kt
@@ -1,4 +1,4 @@
-package com.example.registeractivity
+package com.example.blooddonation.ui.dashboard
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.LocalIndication
@@ -11,13 +11,11 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.unit.dp
