+++ app/src/main/java/com/example/blooddonation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.registeractivity.ui.theme
+package com.example.blooddonation.ui.theme
 
 import androidx.compose.ui.graphics.Color
