+++ app/src/main/java/com/example/blooddonation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.registeractivity.ui.theme
+package com.example.blooddonation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
